// frontend/src/lib/api/model/teamResponseDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Student Arena API
 * API Documentation
 * OpenAPI spec version: 1.0
 */
import type { PlayerResponseDto } from './playerResponseDto';

export interface TeamResponseDto {
  /** Name of the team */
  name: string;
  /** Tag of the team (max 5 characters) */
  tag: string;
  /** description displayed on the team detail page */
  description: string;
  /** Unique identifier for the team */
  id: number;
  /** The user ID of the team owner */
  ownerId: number;
  /** rating for a player */
  rating: number;
  /** The list of team members */
  members: PlayerResponseDto[];
}